pyproject.toml
[project]
name = "yonata"
version = "0.1.3"
description = ""
authors = [
    {name = "Erwin Yonata",email = "erwinwingyonata@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "numpy (>=1.25.2,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "openai (>=1.82.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "deepdiff (>=8.5.0,<9.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
]

[tool.poetry]
packages = [{include = "yonata", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
